+++ examples/root-project/.project/core.json
@@ -18,7 +18,7 @@
   },
   "open_tabs": {
     "tabs": {
-      "04374f11-36e7-409d-827d-593acd7b4cde": {
+      "a8d317ef-2ed8-40b4-8612-0f36b0b1101e": {
         "tab_area": [
           {
             "x": {
@@ -50,24 +50,24 @@
                 101,
                 116,
                 47,
-                100,
+                114,
                 101,
-                98,
-                117,
-                103,
+                108,
+                101,
+                97,
+                115,
+                101,
                 47,
                 102,
-                111,
-                114,
-                116,
-                117,
-                110,
+                105,
+                108,
                 101,
                 95,
-                116,
-                101,
-                108,
-                108,
+                109,
+                97,
+                110,
+                97,
+                103,
                 101,
                 114
               ]
@@ -79,26 +79,26 @@
       }
     },
     "org_tree": {
-      "root_id": "04374f11-36e7-409d-827d-593acd7b4cde",
+      "root_id": "a8d317ef-2ed8-40b4-8612-0f36b0b1101e",
       "nodes": {
-        "04374f11-36e7-409d-827d-593acd7b4cde": {
+        "a8d317ef-2ed8-40b4-8612-0f36b0b1101e": {
           "children": [],
           "parent": null
         }
       }
     },
-    "focused_tab": "04374f11-36e7-409d-827d-593acd7b4cde",
+    "focused_tab": "a8d317ef-2ed8-40b4-8612-0f36b0b1101e",
     "display_tiles": {
-      "root_id": "25fca710-206c-4299-b884-4b79cab41ffc",
+      "root_id": "40a4d34b-16da-4476-91fd-b8f1b90318e1",
       "tiles": {
-        "25fca710-206c-4299-b884-4b79cab41ffc": {
+        "40a4d34b-16da-4476-91fd-b8f1b90318e1": {
           "Tab": {
-            "tab_id": "04374f11-36e7-409d-827d-593acd7b4cde"
+            "tab_id": "a8d317ef-2ed8-40b4-8612-0f36b0b1101e"
           }
         }
       },
       "leaf_registry": {
-        "04374f11-36e7-409d-827d-593acd7b4cde": "25fca710-206c-4299-b884-4b79cab41ffc"
+        "a8d317ef-2ed8-40b4-8612-0f36b0b1101e": "40a4d34b-16da-4476-91fd-b8f1b90318e1"
       }
     }
   }
