+++ singularity_sde/src/project_manager/tabs.rs
@@ -76,7 +76,8 @@ impl Tabs {
             let tabs = Tabs::new_from_root(TabHandler::new(
                 TabData {
                     tab_command: (
-                        "./target/release/fortune_teller".into(),
+                        // "./target/release/fortune_teller".into(),
+                        "./target/release/file_manager".into(),
                         Vec::new(), // Vec::from([""].map(|s| s.into())),
                     ),
                     session_data: serde_json::to_value(project.get_project_directory().clone())
